DBUtils

Description

High level functions for haskell HDBC.

Types

type SQL = String

SQL Statement or SQL expression.

data HDBConn

Database Connection - Objective make the database connection implementation agnostic.

Constructors

HDBConnSqlite Connection

HDBConnPostgres Connection

type DBUriPath = String

Database Uri string.

type SearchIDFun = String -> DBConn [Int]

This type represents a function that takes name and returns all IDs associated with table rows which match the name

type SearchNameIdFun = String -> DBConn [(Int, String)]

type SQLQuery = [(String, SqlValue)]

type DBConn a = forall conn. IConnection conn => ReaderT conn IO a

Database Connection -> DbConn a = ReaderT conn IO a = conn -> IO a

data DBUri

Database URI

Functions

openDBConnection :: DBUriPath -> IO (Maybe HDBConn)

runDBConn :: HDBConn -> DBConn a -> IO a

withDBConnection :: DBUriPath -> DBConn () -> IO ()

withDBConnection2 :: String -> DBConn a -> IO (Maybe a)

withDBConnection3 :: String -> DBConn a -> IO a

withConnection :: IConnection conn => IO conn -> (conn -> IO r) -> IO r

makeSearchIdFun :: SQL -> SearchIDFun

Make function which searches by string or name and returns IDs

makeSearchNameIdFun :: SQL -> SearchNameIdFun

Make function which searches items from table which column matches a string

fromInt64ToInt :: Int64 -> Int

fromIntToInt64 :: Int -> Int64

fromSqlToInt :: SqlValue -> Int

fromSqlToString :: SqlValue -> String

fromIntToHDBC :: Int -> SqlValue

Convert number to HDBC value

fromStrToHDBC :: String -> SqlValue

Convert string to HDBC string

sqlQuery :: SQL -> [SqlValue] -> ([SqlValue] -> b) -> DBConn (Maybe b)

sqlQueryAll :: SQL -> [SqlValue] -> ([SqlValue] -> b) -> DBConn [b]

sqlQueryRow :: SQL -> [SqlValue] -> (SqlValue -> b) -> DBConn [b]

sqlQueryOne :: SQL -> [SqlValue] -> (SqlValue -> b) -> DBConn (Maybe b)

sqlRun :: SQL -> [SqlValue] -> DBConn ()

Run a SQL statement that returns no value

sqlRunMany :: SQL -> [[SqlValue]] -> DBConn ()

sqlCommit :: DBConn () -> DBConn ()

sqlDeleteRowsWhereID :: String -> String -> Int -> DBConn ()

Delete all rows from [table] where [column] = keyID.

Example:

  > sqlDeleteRowsWhereID "itemID" "items" 200

It executes the sql statement:

  > DELETE FROM items WHERE itemID = 200

sqlDeleteRowsTablesWhereID :: String -> [String] -> Int -> DBConn ()

joinStrings :: String -> [String] -> String

Join strings by separator

lookupString :: String -> SQLQuery -> Maybe String

lookupInt :: String -> SQLQuery -> Maybe Int

sqlGetLastID :: String -> String -> DBConn Int