Main

Description

Zotero Web Server and Web API.

Documentation

type ServerApp a = forall conn. IConnection conn => ReaderT conn (ServerPartT IO) a

data ServerConfig

Constructors

ServerConfig

Fields

serverPort :: Int

serverHost :: String

serverStaticPath :: String

serverStoragePath :: String

serverDatabase :: String

serverConf :: Conf

parseInt :: String -> Maybe Int

withConnServerDB :: ToMessage a => String -> Conf -> ServerApp a -> IO ()

withConnServer :: (IConnection conn, ToMessage a) => (String -> IO conn) -> String -> Conf -> ServerApp a -> IO ()

withConnServer2 :: (IConnection conn, ToMessage a) => (String -> IO conn) -> String -> Conf -> ServerApp a -> IO ()

runDbQuery :: DBConn b -> ServerApp b

serverRouteParam :: (String -> Maybe a) -> String -> b -> (a -> DBConn b) -> ServerApp b

serverRouteParamID :: String -> (Int -> DBConn ByteString) -> ServerApp ByteString

Creates a server route for which the paramter is an ID (int) - Identification Number for the database query

serverRouteParamString :: String -> (String -> DBConn ByteString) -> ServerApp ByteString

Creates a server route for which the paramter is an ID - Identification Number for the database query

makeRoutes :: String -> String -> ServerApp Response

parseDbDriver :: String -> String

runServerConf :: ServerConfig -> IO ()

runServer :: t -> String -> String -> String -> String -> IO ()

loadServerConf :: FilePath -> IO ()

routeCollection :: ServerApp ByteString

Route that displays all collections Rest API: http://localhost:8000/api/colls

routeCollectionID :: ServerApp ByteString

Show all items from a given collection defined by its ID Rest API - apicolls?id=23423 or apicoll?id={collection ID}

routeTagID :: ServerApp ByteString

Returns all items from a given tagID Rest API - http://localhost:8000/api/tags?id=15 or http://localhost:8000/api/tags?id=[tagID]

routeTags :: ServerApp ByteString

Returns all tags Rest API - http://localhost:8000/api/tags

routeAuthorID :: ServerApp ByteString

Get all items belonging to an author, given its ID Rest API - http://localhost:8000/api/authors?id=100

routeAuthors :: ServerApp ByteString

Return all authors data Rest API - http://localhost:8000/api/authors

routeTagsFromCollID :: ServerApp ByteString

routeSearchByTitleLike :: ServerApp ByteString

routeRealatedTags :: ServerApp ByteString

routeSearchByContentAndTitleLike :: ServerApp ByteString

routeItemsWithoutCollection :: ServerApp ByteString

Server http request with all items without collections.

routeItemsWithoutCollection2 :: ServerApp ByteString

showUserHelp :: IO ()

Show user help command

parseArgs :: [String] -> IO ()

Handle command line arguments

main :: IO ()

Main IO Action