Zotero

Description

Interface to Zotero database to query and manipulate the database.

Types

type DBConn a = forall conn. IConnection conn => ReaderT conn IO a

Database Connection -> DbConn a = ReaderT conn IO a = conn -> IO a

data ZoteroItem

ZoteroItem data

Constructors

ZoteroItem

Fields

zoteroItemID :: Int

zoteroItemData :: [(String, String)]

zoteroItemAuthors :: [ZoteroAuthor]

zoteroItemTags :: [(Int, String)]

zoteroItemCollections :: [(Int, String)]

zoteroItemFile :: Maybe String

zoteroItemMime :: Maybe String

data ZoteroAuthor

Constructors

ZoteroAuthor

Fields

zoteroAuthorID :: Int

zoteroAuthorFirstName :: String

zoteroAuthorLastName :: String

data ZoteroTag

Constructors

ZoteroTag

Fields

zoteroTagID :: Int

zoteroTagName :: String

data ZoteroColl

Constructors

ZoteroColl

Fields

zoteroCollID :: Int

zoteroCollName :: String

type ZoteroItemID = Int

type ZoteroTagID = Int

Zotero Tag ID number

type ZoteroTagName = String

Tag Name

type ZoteroItemTags = [(ZoteroTagID, ZoteroTagName)]

type ZoteroItemMime = String

Functions

withConnection :: IConnection conn => IO conn -> (conn -> IO r) -> IO r

getCollections :: DBConn [ZoteroColl]

Get all collections

collectionItems :: Int -> DBConn [Int]

itemTagsData :: ZoteroItemID -> DBConn ZoteroItemTags

Returns all tags of a given item

itemTags :: ZoteroItemID -> DBConn [String]

Return only the tag names of a given itemID

itemData :: ZoteroItemID -> DBConn [(String, String)]

itemAttachmentData :: ZoteroItemID -> DBConn (Maybe [String])

Get Zotero item attachment file.

itemAttachmentFile :: Int -> DBConn (Maybe FilePath)

itemAuthors :: ZoteroItemID -> DBConn [ZoteroAuthor]

sqlQuery :: String -> [SqlValue] -> ([SqlValue] -> b) -> DBConn (Maybe b)

sqlQueryAll :: String -> [SqlValue] -> ([SqlValue] -> b) -> DBConn [b]

sqlQueryRow :: String -> [SqlValue] -> (SqlValue -> b) -> DBConn [b]

dbConnection :: IO Connection

getTagItems :: Int -> DBConn [ZoteroItemID]

Get all items IDs that belong to a given tag specified by its Id

getRelatedTags :: ZoteroTagID -> DBConn [ZoteroTag]

getTagsFromCollection :: Int -> DBConn [ZoteroTag]

searchByTitleWordLike :: String -> DBConn [Int]

searchByContentAndTitleLike :: String -> DBConn [Int]

getCollectionChild :: Int -> DBConn [ZoteroColl]

Get sub-collections of a collection

getCollectionTop :: DBConn [ZoteroColl]

Get only top level collections

itemsWithoutCollections :: Int -> Int -> DBConn [ZoteroItemID]

Return all zotero items without collections

getCollectionsJSON :: DBConn ByteString

Get all collections as JSON

getTagsJSON :: DBConn ByteString

getZoteroItem :: ZoteroItemID -> DBConn ZoteroItem

getZoteroItem - Get Zotero Item from the database

getZoteroItemJSON :: ZoteroItemID -> DBConn ByteString

getZoteroItemsJSON :: [ZoteroItemID] -> DBConn ByteString

Get a list of zotero item data as json given its IDs

getCollectionItemsJSON :: Int -> DBConn ByteString

getTagItemsJSON :: ZoteroTagID -> DBConn ByteString

getItemsFromAuthorJSON :: Int -> DBConn ByteString

getAuthorsJSON :: DBConn ByteString

getRelatedTagsJSON :: ZoteroTagID -> DBConn ByteString

getTagsFromCollectionJSON :: Int -> DBConn ByteString

searchByTitleWordLikeJSON :: String -> DBConn ByteString

searchByContentAndTitleLikeJSON :: String -> DBConn ByteString

getCollectionChildJSON :: Int -> DBConn ByteString

getCollectionTopJSON :: DBConn ByteString

itemsWithoutCollectionsJSON :: Int -> Int -> DBConn ByteString

getZoteroItemIdAsListJSON :: ZoteroItemID -> DBConn ByteString

getItemsFromAuthor :: Int -> DBConn [Int]

getAuthors :: DBConn [ZoteroAuthor]

joinStrings :: String -> [String] -> String

strip :: String -> String